Print PbGen logs
