+++ src/cpp/benders/benders_core/include/core/ILogger.h
@@ -1,12 +1,9 @@
 #ifndef ANTARESXPANSION_ILOGGER_H
 #define ANTARESXPANSION_ILOGGER_H
 
-#include <time.h>
-
 #include <filesystem>
 #include <map>
 #include <memory>
-#include <sstream>
 #include <string>
 #include <vector>
 
@@ -78,36 +75,6 @@ struct LogData {
            subproblem_time == lhs.subproblem_time;
   }
 };
-inline std::string format_time_str(const long int time_in_seconds) {
-  std::time_t seconds(
-      time_in_seconds);  // you have to convert your input_seconds into time_t
-  std::tm p;
-
-  // convert to broken down time
-#if defined(WIN32) || defined(_WIN32) || defined(__WIN32__) || defined(__NT__)
-  gmtime_s(&p, &seconds);
-#else  // defined(__unix__) || (__APPLE__)
-  gmtime_r(&seconds, &p);
-#endif
-  std::stringstream ss;
-  const auto days = p.tm_yday;
-  const auto hours = p.tm_hour;
-  const auto mins = p.tm_min;
-  const auto secs = p.tm_sec;
-  if (days > 0) {
-    ss << days << " days ";
-  }
-  if (hours > 0) {
-    ss << hours << " hours ";
-  }
-  if (mins > 0) {
-    ss << mins << " minutes ";
-  }
-  if (secs > 0) {
-    ss << secs << " seconds ";
-  }
-  return ss.str();
-}
 class ILogger {
  public:
   virtual ~ILogger() = default;
