+++ src/cpp/benders/logger/User.cpp
@@ -9,6 +9,7 @@
 #include "CandidateLog.h"
 #include "Commons.h"
 #include "IterationResultLog.h"
+#include "Timer.h"
 
 using xpansion::logger::commons::indent_1;
 namespace xpansion {
